refactor: move requestIdleCallbackPolyfill to utils directory
Extract requestIdleCallbackPolyfill and related functions into a dedicated browser-polyfills utility file for better code organization and reusability.
